layout/base/tests/test_bug970964.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=970964
-->
<head>
  <title>Test for Bug 970964</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <script type="text/javascript">
    function testWithoutImplicitPointerCapture() {
      var iframe = document.getElementById("testFrame");
      iframe.src = "bug970964_inner.html";
    }

    function testWithImplicitPointerCapture() {
      var iframe = document.getElementById("testFrame");
      iframe.src = "bug970964_inner2.html";
    }

    function runTest() {
      SimpleTest.waitForExplicitFinish();
      window.addEventListener("message", (event) => {
        if (event.data == "finishTest") {
          SimpleTest.finish();
        } else if (event.data == "run next") {
          SpecialPowers.pushPrefEnv({
            "set": [
              ["dom.w3c_pointer_events.enabled", true],
              ["dom.w3c_pointer_events.implicit_capture", true]
            ]
          }, testWithImplicitPointerCapture);
        }
      });

      SpecialPowers.pushPrefEnv({
        "set": [
          ["dom.w3c_pointer_events.enabled", true],
          ["dom.w3c_pointer_events.implicit_capture", false]
        ]
      }, testWithoutImplicitPointerCapture);
    }
  </script>
</head>
<body onload="runTest();">
  <iframe id="testFrame" height="500" width="500"></iframe>
</body>
